🛠💣 major refactoring: introduce scoring module
- fixes #18
- the module can be used to integrate scoring feature for fuzzywuzzy and regular expressions
