+++ nlpurify/scoring/__init__.py
@@ -1,0 +1,11 @@
+# -*- encoding: utf-8 -*-
+
+"""
+The module provides scoring methods for text matching using different
+modules like fuzzywuzzy, regular expressions, etc.
+"""
+
+from nlpurify.scoring.fuzzy import (
+    fuzzy_score,
+    LogicalFuzzy
+)
